Organize link analysis and isomorphism
